Move data structs to model.go.
